requirements.txt
asgiref==3.8.1
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
Markdown==3.7
packaging==24.1
pillow==10.4.0
python-decouple==3.8
sqlparse==0.5.1
tzdata==2024.1
whitenoise==6.7.0
